Adds functions for Polish dictionaries
Added functions for Polish wordreference (okay but not great support because
not enough words and can't handle any level of inflection really). Added a
function for the lektorek website, which is a great website, but unfortunately
so far cannot get it to work, as neither direct calls to the API, nor embedding
in iFrame and selecting by ID to enter the search term and submit will work.
This is due to different origin policies on chrome.