Implemented lektorek Polish dictionary
Added support for Lektorek Polish dictionary. This requires Chrome to
be in a mode with CORS disabled - I use CORS toggle add-on for Chrome.
I'm pulling json data from the lektorek api from which I then display
the 'embedded_definition' (this already has html markup for bold and
italic classes which I added to the css. I've styled it so that as new
searches are made by clicking, the new result is added as a new DIV at
the start of the list of DIVs with previous searches moving down and
getting gray text through a CSS class.